+++ docs/source/changelog.rst
@@ -5,6 +5,7 @@ Change Log
 **Pyro 4.50**
 
 - Future can now be given a delay before it is evaluated
+- Future can now be cancelled (if it hasn't been evalued yet)
 - new ITER_STREAM_LINGER config item to keep streams alive for a given period after proxy disconnect (defaults to 30 sec.)
 - new NS_AUTOCLEAN config item to set a recurring period in seconds where the Name server checks its registrations.
   It will then auto cleanup registrations after a short while if they're no longer available. (defaults to 0.0 - disabled).